# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/instructors": {
      "post": {
        "operationId": "InstructorsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInstructorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Create a new instructor",
        "tags": [
          "Instructors"
        ]
      },
      "get": {
        "operationId": "InstructorsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all instructors",
        "tags": [
          "Instructors"
        ]
      }
    },
    "/instructors/{id}": {
      "get": {
        "operationId": "InstructorsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get an instructor by id",
        "tags": [
          "Instructors"
        ]
      },
      "patch": {
        "operationId": "InstructorsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInstructorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update an instructor",
        "tags": [
          "Instructors"
        ]
      },
      "delete": {
        "operationId": "InstructorsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete an instructor",
        "tags": [
          "Instructors"
        ]
      }
    },
...
    "/learners": {
      "get": {
        "operationId": "LearnersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all learners",
        "tags": [
          "Learners"
        ]
      },
      "post": {
        "operationId": "LearnersController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Create a new learner",
        "tags": [
          "Learners"
        ]
      }
    },
    "/learners/{id}": {
      "get": {
        "operationId": "LearnersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get learner by ID",
        "tags": [
          "Learners"
        ]
      },
      "patch": {
        "operationId": "LearnersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update learner",
        "tags": [
          "Learners"
        ]
      },
      "delete": {
        "operationId": "LearnersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete learner",
        "tags": [
          "Learners"
        ]
      }
    },
    "/organizations": {
      "get": {
        "operationId": "OrganizationsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all organizations",
        "tags": [
          "Organizations"
        ]
      },
      "post": {
        "operationId": "OrganizationsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Create a new organization",
        "tags": [
          "Organizations"
        ]
      }
    },
    "/organizations/{id}": {
      "get": {
        "operationId": "OrganizationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get organization by ID",
        "tags": [
          "Organizations"
        ]
      },
      "patch": {
        "operationId": "OrganizationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update organization",
        "tags": [
          "Organizations"
        ]
      },
      "delete": {
        "operationId": "OrganizationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete organization",
        "tags": [
          "Organizations"
        ]
      }
    },
...
    "/notifications": {
      "get": {
        "operationId": "NotificationsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all notifications",
        "tags": [
          "Notifications"
        ]
      },
      "post": {
        "operationId": "NotificationsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Create new notification",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/{id}": {
      "get": {
        "operationId": "NotificationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get notification by ID",
        "tags": [
          "Notifications"
        ]
      },
      "patch": {
        "operationId": "NotificationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update notification status",
        "tags": [
          "Notifications"
        ]
      },
      "delete": {
        "operationId": "NotificationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete notification",
        "tags": [
          "Notifications"
        ]
      }
    },
...
    "/refunds": {
      "get": {
        "operationId": "RefundsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all refund requests",
        "tags": [
          "Refunds"
        ]
      }
    },
    "/refunds/{id}": {
      "get": {
        "operationId": "RefundsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get refund request by ID",
        "tags": [
          "Refunds"
        ]
      },
      "patch": {
        "operationId": "RefundsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Approve or reject a refund",
        "tags": [
          "Refunds"
        ]
      }
    },
...
    "/quizzes": {
      "get": {
        "operationId": "QuizzesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all quizzes",
        "tags": [
          "Quizzes"
        ]
      },
      "post": {
        "operationId": "QuizzesController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Create new quiz",
        "tags": [
          "Quizzes"
        ]
      }
    },
    "/quizzes/{id}": {
      "get": {
        "operationId": "QuizzesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get quiz by ID",
        "tags": [
          "Quizzes"
        ]
      },
      "patch": {
        "operationId": "QuizzesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update quiz",
        "tags": [
          "Quizzes"
        ]
      }
    }
  },
  "info": {
    "title": "Coursera App API",
    "description": "The API for Admin and Instructor modules",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role"
      }
    },
    "schemas": {
      "CreateInstructorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the instructor",
            "example": "Jane Smith"
          },
          "email": {
            "type": "string",
            "description": "The email of the instructor",
            "example": "instructor@example.com"
          },
          "expertise": {
            "type": "string",
            "description": "The expertise of the instructor",
            "example": "Machine Learning"
          },
          "qualification": {
            "type": "string",
            "description": "The qualification of the instructor",
            "example": "Ph.D. in AI"
          },
          "organization": {
            "type": "string",
            "description": "The organization of the instructor",
            "example": "Tech Academy"
          },
          "verificationStatus": {
            "type": "string",
            "description": "The verification status",
            "example": "Verified"
          },
          "status": {
            "type": "string",
            "description": "The status",
            "example": "Active"
          },
          "courses": {
            "type": "number",
            "description": "Number of courses",
            "example": 5
          },
          "phone": {
            "type": "string",
            "description": "Phone number"
          },
          "location": {
            "type": "string",
            "description": "Location"
          },
          "website": {
            "type": "string",
            "description": "Personal website URL"
          },
          "bio": {
            "type": "string",
            "description": "Instructor bio"
          },
          "orgDepartment": {
            "type": "string",
            "description": "Organization department"
          },
          "orgRole": {
            "type": "string",
            "description": "Organization role"
          },
          "orgWorkEmail": {
            "type": "string",
            "description": "Institutional email"
          },
          "orgWebsite": {
            "type": "string",
            "description": "Organization website"
          },
          "orgJoined": {
            "type": "string",
            "description": "Organization joined date"
          }
        },
        "required": [
          "name",
          "email"
        ]
      },
      "UpdateInstructorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the instructor",
            "example": "Jane Smith"
          },
          "email": {
            "type": "string",
            "description": "The email of the instructor",
            "example": "instructor@example.com"
          },
          "expertise": {
            "type": "string",
            "description": "The expertise of the instructor",
            "example": "Machine Learning"
          },
          "qualification": {
            "type": "string",
            "description": "The qualification of the instructor",
            "example": "Ph.D. in AI"
          },
          "organization": {
            "type": "string",
            "description": "The organization of the instructor",
            "example": "Tech Academy"
          },
          "verificationStatus": {
            "type": "string",
            "description": "The verification status",
            "example": "Verified"
          },
          "status": {
            "type": "string",
            "description": "The status",
            "example": "Active"
          },
          "courses": {
            "type": "number",
            "description": "Number of courses",
            "example": 5
          },
          "phone": {
            "type": "string",
            "description": "Phone number"
          },
          "location": {
            "type": "string",
            "description": "Location"
          },
          "website": {
            "type": "string",
            "description": "Personal website URL"
          },
          "bio": {
            "type": "string",
            "description": "Instructor bio"
          },
          "orgDepartment": {
            "type": "string",
            "description": "Organization department"
          },
          "orgRole": {
            "type": "string",
            "description": "Organization role"
          },
          "orgWorkEmail": {
            "type": "string",
            "description": "Institutional email"
          },
          "orgWebsite": {
            "type": "string",
            "description": "Organization website"
          },
          "orgJoined": {
            "type": "string",
            "description": "Organization joined date"
          }
        }
      },
...
    }
  }
}